requirements.txt
psycopg2-binary==2.9.3
dict2xml==1.7.0
urllib3==1.26.8